JBIDE-22321: extract class collecting behaviour
